+++ data/sae.txt
@@ -1,3 +1,4 @@
 D | 0 | return book | 2 December 2019 6pm
 T | 0 | sleep
 E | 1 | sleep | 2pm | 3pm
+T | 0 | sleep
