+++ src/main/java/sae/exceptions/InvalidDateTimeFormatException.java
@@ -1,3 +1,5 @@
+package sae.exceptions;
+
 public class InvalidDateTimeFormatException extends Exception {
     public InvalidDateTimeFormatException(String message) {
         super(message);
