+++ src/main/java/sae/exceptions/InvalidEventException.java
@@ -1,3 +1,5 @@
+package sae.exceptions;
+
 public class InvalidEventException extends SaeException {
 
     @Override
