+++ src/main/java/sae/task/Deadline.java
@@ -1,24 +1,21 @@
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
+package sae.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 
 /**
- * The Deadline class represents a task with a specific deadline.
- * It extends the Task class and adds a 'by' field to store the deadline.
+ * The Deadline class represents a sae.task with a specific deadline.
+ * It extends the sae.task.Task class and adds a 'by' field to store the deadline.
  */
 public class Deadline extends Task {
     protected LocalDateTime by;
 
     /**
-     * Constructs a new Deadline task with the given description and deadline.
+     * Constructs a new Deadline sae.task with the given description and deadline.
      *
-     * @param description The description of the task.
-     * @param by          The deadline of the task.
+     * @param description The description of the sae.task.
+     * @param by          The deadline of the sae.task.
      */
     public Deadline(String description, LocalDateTime by) {
         super(description);
@@ -26,9 +23,9 @@ public class Deadline extends Task {
     }
 
     /**
-     * Overrides the toString method to format the Deadline task's details.
+     * Overrides the toString method to format the Deadline sae.task's details.
      *
-     * @return A formatted string representing the Deadline task.
+     * @return A formatted string representing the Deadline sae.task.
      */
     @Override
     public String toString() {
