+++ src/main/java/sae/task/Event.java
@@ -1,11 +1,13 @@
+package sae.task;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 
 /**
- * The Event class represents a task that occurs within a specific time frame.
- * It extends the Task class and adds 'from' and 'to' fields to store the event timing.
+ * The sae.task.Event class represents a sae.task that occurs within a specific time frame.
+ * It extends the sae.task.Task class and adds 'from' and 'to' fields to store the event timing.
  */
 public class Event extends Task {
 
@@ -13,9 +15,9 @@ public class Event extends Task {
     protected String to;
 
     /**
-     * Constructs a new Event task with the given description, start time, and end time.
+     * Constructs a new sae.task.Event sae.task with the given description, start time, and end time.
      *
-     * @param description The description of the task.
+     * @param description The description of the sae.task.
      * @param from        The start time of the event.
      * @param to          The end time of the event.
      */
@@ -26,9 +28,9 @@ public class Event extends Task {
     }
 
     /**
-     * Overrides the toString method to format the Event task's details.
+     * Overrides the toString method to format the sae.task.Event sae.task's details.
      *
-     * @return A formatted string representing the Event task.
+     * @return A formatted string representing the sae.task.Event sae.task.
      */
     @Override
     public String toString() {
