+++ text-ui-test/EXPECTED.TXT
@@ -1,12 +1,12 @@
 Hello! I'm Sae
 What can I do for you?
-Got it. I've added this task:
+Got it. I've added this sae.task:
 [T][ ] sleep
 Now you have 1 tasks in the list.
-Got it. I've added this task:
+Got it. I've added this sae.task:
 [D][ ] cs2103T ip (by: thurs 2359)
 Now you have 2 tasks in the list.
-Got it. I've added this task:
+Got it. I've added this sae.task:
   [E][ ] lak class (from: thur 6pm to: thur 8pm)
 Now you have 3 tasks in the list.
 Bye. Hope to see you again soon!
